refactor: Split summary and info for alert configs
